move merkleproof to native
